% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiasCorrectionE.R
\name{BiasCorrectionE}
\alias{BiasCorrectionE}
\title{BiasCorrectionE}
\usage{
BiasCorrectionE(D, nf, ncl, para)
}
\arguments{
\item{D}{The data frame. The first 'nf' columns are the covariates under study. Other columns must include fields
'time' for time-to-failure, 'Cens' for censoring and 'cluster' for cluster identification. The number of rows is equal to the number of subjects.}

\item{nf}{The number of covariates under study.}

\item{ncl}{The number of clusters.}

\item{para}{The vector of the ML estimates of unknown parameters. This vector has a form
\deqn{y = (ln(a), \beta , ln(\sigma ^2)),}
where \eqn{a} is the slope, \eqn{\beta } is the vector of the Cox-regression parameters,
and \eqn{\sigma ^2} is the variance of frailty.}
}
\value{
The bias correction for vector \eqn{y}. If the calculation of the bias correction is
not possible returns NAs.
}
\description{
Returns the bias correction for ML estimates.
}
\details{
This function calculates the bias correction for the proportional hazards model with gamma frailty
in accordance with the Cox-Snell method.  The baseline cumulative hazard function is \deqn{H_{0}(t;a,b)=at}.
}
\examples{
\dontrun{
BiasCorrectionE(D,nf,ncl,para)

}
}
