% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiasCorrectionW.R
\name{BiasCorrectionW}
\alias{BiasCorrectionW}
\title{BiasCorrectionW}
\usage{
BiasCorrectionW(D, nf, ncl, para)
}
\arguments{
\item{D}{The data frame. The first 'nf' columns are the covariates under study. Other columns must include fields
'time' for time-to-failure, 'Cens' for censoring and 'cluster' for cluster identification. The number of rows is equal to the number of subjects.}

\item{nf}{The number of covariates under study.}

\item{ncl}{The number of clusters.}

\item{para}{The vector of the ML estimates of unknown parameters. This vector has a form \deqn{y = (ln(a), \beta , ln(\sigma ^2), b),}
where \eqn{a} and \eqn{b} are the slope and the shape parameters, \eqn{\beta } is the vector of the Cox-regression parameters,
and \eqn{\sigma ^2} is the variance of the frailty.}
}
\value{
The bias correction for vector \eqn{y}. If the calculation of the bias correction is
not possible returns NAs.
}
\description{
Returns the bias correction for ML estimates.
}
\details{
This function calculates the bias correction for the proportianal hazards model with gamma frailty
in accordance with the Cox-Snell method. The baseline cumulative hazard function is \deqn{H_{0}(t;a,b)=at^b}.
}
\examples{
\dontrun{
BiasCorrectionW(D,nf,ncl,para)

}
}
