% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{change}
\alias{change}
\title{change}
\usage{
change(Text, i1, i2, i3)
}
\arguments{
\item{Text}{A character matrix, where 2-10 columns are numbers in character format.}

\item{i1}{A number.}

\item{i2}{A number.}

\item{i3}{A number.}
}
\value{
Changed character matrix 'Text'.
}
\description{
Adds values of i1, i2 and i3 to 8th, 9th and 10th columns of character matrix 'Text'.
}
\examples{
\dontrun{
change(Text,i1,i2,i3)
}

}
