% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{convert}
\usage{
convert(Text, i1, i2, Parnam, Parmen)
}
\arguments{
\item{Text}{A character matrix, where the first column contains expressions in text format.
Other 9 columns contain non-negative integer numbers.}

\item{i1}{first index.}

\item{i2}{second index.}

\item{Parnam}{The vector of the names of parameters needed to calculate expressions.}

\item{Parmen}{The vector of the values of parameters needed to calculate expressions.}
}
\value{
The sum of values over all expressions in the first column 'Text'.
}
\description{
Calculates the sum of values over expressions.
}
\details{
For each row of the matrix 'Text' this function calculates the value of the following expression:
\deqn{Expr=n^Deg*T_1*c_1^(T_2*ND1+T_5)*la1_i1_(T_2*ND1+T_5)_T8*c_2^(T_3*ND2+T_6)*la2_i2_(T_3*ND1+T_6)_T9*c_3^(T_4*ND1+T_7)*la3_0_(T_4*ND1+T_7)_T10}
and then takes the sum over all rows. Here \eqn{T_1} (first column in parameter 'Text') is a character string standing for an expression,
\eqn{T_1}, \eqn{i=2,...,10} (last 9 columns in parameter 'Table') are the numbers in character format,
other variables (with exception of \eqn{i1} and \eqn{i2}) are the global parameters.
}
\examples{
\dontrun{
convert(Text,i1,i2)
}

}
