% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diff.R
\name{Diff}
\alias{Diff}
\title{Diff}
\usage{
Diff(Text, x, order = 1)
}
\arguments{
\item{Text}{A character matrix, where the first column contains expressions in text format that must be differentiated.
Other 9 columns contain non-negative integer numbers.}

\item{x}{character vector, giving the variable name including numerical index (e.g. 'x1') with respect to which derivatives will be computed.}

\item{order}{Order of differentiation. A positive integer number.}
}
\value{
Changed character matrix 'Text'.
}
\description{
Differentiates and modifies (if necessary) an expression given in variable 'Text'.
}
\details{
This function is written specially for this package and takes into account the special structure of matrix 'Text'.
Modification of the matrix 'Text' depends on the content of the columns 2-10 of matrix 'Text' and on the index part of variable 'x'.
}
\examples{
\dontrun{
Text=matrix(c("x1^2","x1^3",rep(c("2","1"),9)),2,10)
Diff(Text,"x1",1)
}

}
