% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diff0.R
\name{Diff0}
\alias{Diff0}
\title{Diff0}
\usage{
Diff0(Text, x)
}
\arguments{
\item{Text}{A character matrix, where the first column contains expressions in text format that must be differentiated.
Other 9 columns contain non-negative integer numbers.}

\item{x}{A character vector, giving the variable name  with respect to which derivatives will be computed.}
}
\value{
Diff0 return a call.
}
\description{
Provides derivatives of first order.
}
\examples{
\dontrun{
Text=matrix(c("x^2","x^3",rep(c("2","1"),9)),2,10)
Diff0(Text,"x")
}

}
