% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpg.R
\name{dpg}
\alias{dpg}
\title{dpg}
\usage{
dpg(t)
}
\arguments{
\item{t}{An expression.}
}
\value{
Parsed object.
}
\description{
Returns the parsed but unevaluted expression.
}
\details{
Convert expression 't' into parsed object.
}
\examples{
\dontrun{
t=expression(a+b*x)
dpg(t) 
}

}
