% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikMarg.R
\name{LikMarg}
\alias{LikMarg}
\title{LikMarg}
\usage{
LikMarg(y, D, nf, ncl, dist)
}
\arguments{
\item{y}{Vector of parameters in the form:
\deqn{y = (ln(a), \beta , ln(\sigma ^2))}
for the exponential baseline hazard function,
\deqn{y = (ln(a), \beta , ln(\sigma ^2),b)}
for the Weibull baseline hazard function and
\deqn{y = (ln(10^3a), \beta , ln(\sigma ^2), 10b)}
for the Gompertz baseline hazard function, where a and b are the slope and the shape parameters,
\eqn{\beta } are the Cox-regression parameters, and \eqn{\sigma ^2} is the variance of frailty.}

\item{D}{A data.frame in which to interpret the variables named in the formula.
The data set includes the following fields:
\enumerate{
\item time-to-failure 'time';
\item censoring 'event' (censoring must be either 0 for no event or 1 for event);
\item  covariates (continuous or categorical, first 'nf' columns) used in a study (can be empty set);
\item cluster - the name of a cluster variable in data (is equal to NULL for the fixed-effect model).}}

\item{nf}{The number of covariates under study.}

\item{ncl}{The number of clusters.}

\item{dist}{Baseline hazard function ('Exponential', 'Weibull' or 'Gompertz').}
}
\value{
Neg-loglikelihood
}
\description{
This function calculates neg-loglikelihood for a proportional hazards model.
}
\details{
Three kinds of the baseline hazards are used in this function:
\enumerate{
\item Exponential with baseline cumulative hazard function \deqn{H_{0}(t;a)=at;}
\item Weibull with baseline cumulative hazard function \deqn{H_{0}(t;a,b)=at^b;}
\item Gompertz with baseline cumulative hazard function \deqn{H_{0}(t;a,b)=a(exp(bt)-1)).}
}
}
\examples{
\dontrun{
LikMarg(y,D,nf,ncl,dist)
}
}
