% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.model.R
\name{opt.model}
\alias{opt.model}
\title{opt.model}
\usage{
opt.model(formula, data, cluster)
}
\arguments{
\item{formula}{A formula object, with the response on the left of a ~ operator, and the terms on
the right. The response must be a survival object as returned by the Surv() function.
The status indicator corresponding to censoring in the Surv object must be 0 for censored objects and 1 for non-censored ones.
This object describes the effect of several factors in the Cox proportional hazards model.}

\item{data}{The data frame. Includes fields for:
\enumerate{
\item Time-to-failure;
\item Censoring;
\item Possible covariates (optional);
\item Cluster (if any).
}
The names of these fields must be included in 'formula'.}

\item{cluster}{The name of a cluster variable in data (is equal to NULL for the fixed-effect model).}
}
\value{
Retutns the latex table with AIC and BIC values for three hazard functions: exponential, Gompertz and Weibull.
}
\description{
Calculates AIC and BIC for ML estimator applied to the proportional hazards model with gamma frailty
and different hazard functions.
}
\examples{
\dontrun{
library(survival)
D=veteran
D$prior <- factor(as.character(D$prior), labels = c(0, 1))
D$trt <- factor(as.character(D$trt), labels = c(0, 1))
cluster="cluster"
D=data.frame(D$karno,D$trt,D$diagtime,D$prior,D$age,D$celltype,D$time,D$status)
colnames(D)=c("karno","trt","dtime","prior","age","cluster","time","status")
dist="Exponential"
formula=as.formula("Surv(time, status) ~ karno")
RRR=opt.model(formula,D,cluster)
print(RRR)
}

}
