% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsi1c.R
\name{subsi1c}
\alias{subsi1c}
\title{subsi1c}
\usage{
subsi1c(Zag, nf, Parnam, Parmen, WW6)
}
\arguments{
\item{Zag}{A character object. Includes variables \eqn{x_i} corresponding to covariates.}

\item{nf}{The number of covariates.}

\item{Parnam}{The vector of the names of parameters needed to calculate expression.}

\item{Parmen}{The vector of the values of parameters needed to calculate expression.}

\item{WW6}{The matrix of covariates.}
}
\value{
The full sum of an expression over all subjects
}
\description{
Calculates the sum of elements for the matrix of expressions.
}
\details{
Text 'Zag' stands for an expression. Variables \eqn{x_i} are replaced
with respective covariates and the full sum over all subjects is calculated. This function is used in
function 'BiasCorrectionG' only.
}
\examples{
\dontrun{
subsi1c(Zag,nf,Parnam,Parmen,WW6)
}

}
