% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsi2c.R
\name{subsi2c}
\alias{subsi2c}
\title{subsi2c}
\usage{
subsi2c(Zag, nf, Parnam, Parmen, WW4, WW5)
}
\arguments{
\item{Zag}{A character object. Includes variables \eqn{x_i} and \eqn{x__i} corresponding to covariates.}

\item{nf}{The number of covariates.}

\item{Parnam}{The vector of the names of parameters needed to calculate expression.}

\item{Parmen}{The vector of the values of parameters needed to calculate expression.}

\item{WW4}{The vector of the matrices of covariates. Each such matrix is the \eqn{n}-times repeated vector-column
standing for a covariate, where \eqn{n} is the number of subjects.}

\item{WW5}{The vector of the transposed matrices of covariates.}
}
\value{
The full sum of expression over all subjects
}
\description{
Calculates the sum of elements for the matrix of expressions.
}
\details{
Text 'Zag' stands for an expression. Variables \eqn{x_i} and \eqn{x__i} are replaced
with matrix of covariates and its transposed, respectively. Then the full sum of expression over all subjects is calculated.
This function is used in function 'BiasCorrectionG' only.
}
\examples{
\dontrun{
subsi2c(Zag,nf,Parnam,Parmen,WW4,WW5)
}

}
