% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsi2c_i.R
\name{subsi2c_i}
\alias{subsi2c_i}
\title{subsi2c}
\usage{
subsi2c_i(Zag, nf, Parnam, Parmen, WW6)
}
\arguments{
\item{Zag}{A character object. Includes variables \eqn{x_i} and \eqn{x__i} corresponding to covariates.}

\item{nf}{The number of covariates.}

\item{Parnam}{The vector of the names of parameters needed to calculate expression.}

\item{Parmen}{The vector of the values of parameters needed to calculate expression.}

\item{WW6}{The matrix of covariates.}
}
\value{
The full sum of expression over all subjects
}
\description{
Calculates the sum of elements for the matrix of expressions.
}
\details{
Text 'Zag' is a character representation of an expression. Variables \eqn{x_i} and \eqn{x__i} are replaced
with respective matrices of covariates and the full sum of expression over all subjects is calculated.
This function is used in function 'BiasCorrectionG' only.
}
\examples{
\dontrun{
subsi2c_i(Zag,nf,Parnam,Parmen,WW6)
}

}
