% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsi3c_i123.R
\name{subsi3c_i123}
\alias{subsi3c_i123}
\title{subsi3c_i13}
\usage{
subsi3c_i123(Zag, nf, Parnam, Parmen, WW6)
}
\arguments{
\item{Zag}{A character object. Includes variables 'x_i', 'x__i' and 'x___i' corresponding to covariates.}

\item{nf}{The number of covariates.}

\item{Parnam}{The vector of the names of parameters needed to calculate expression.}

\item{Parmen}{The vector of the values of parameters needed to calculate expression.}

\item{WW6}{The matrix of covariates.}
}
\value{
The full sum of expression over all subjects
}
\description{
Calculates the full sum of elements for the matrix of expressions.
}
\details{
Text 'Zag' stands for an expression. Variables \eqn{x_i}, \eqn{x__i} and \eqn{x___i} are replaced
with matrices of covariates Finally, the full sum of expression over all subjects is calculated.
This function is used in function 'BiasCorrectionG' only.
}
\examples{
\dontrun{
subsi3c_i123(Zag,nf,Parnam,Parmen,WW6)
}

}
