% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsi3c_i13.R
\name{subsi3c_i13}
\alias{subsi3c_i13}
\title{subsi3c_i13}
\usage{
subsi3c_i13(Zag, nf, Parnam, Parmen, WW4, WW5)
}
\arguments{
\item{Zag}{A character object. Includes variables \eqn{x_i}, \eqn{x__i} and \eqn{x___i} corresponding to covariates.}

\item{nf}{The number of covariates.}

\item{Parnam}{The vector of the names of parameters needed to calculate expression.}

\item{Parmen}{The vector of the values of parameters needed to calculate expression.}

\item{WW4}{The vector of the matrices of covariates. Each such matrix is the \eqn{n}-times repeated vector-column
standing for a covariate, where \eqn{n} is the number of subjects.}

\item{WW5}{The vector of the transposed matrices of covariates.}
}
\value{
The full sum of expression over all subjects
}
\description{
Calculates the full sum of elements for the matrix of expressions.
}
\details{
Text 'Zag' stands for an expression. Variables \eqn{x_i}, \eqn{x___i} and \eqn{x__i} are replaced
with matrices of covariates for \eqn{x_i} and \eqn{x___i} and with matrix of transposed for \eqn{x__i}.
Finally, the full sum of expression over all subjects is calculated.
This function is used in function 'BiasCorrectionG' only.
}
\examples{
\dontrun{
subsi3c_i13(Zag,nf,Parnam,Parmen,WW4,WW5)
}

}
